Merge from dev to main

+++ src/main/java/com/amitit/webapp/entity/Course.java
@@ -19,22 +19,22 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Course {
-	
+
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long cid;
-	
+	private int cid;
+
 	private String name;
 	private String description;
 	private String duration;
-	
+
 	@OneToOne(cascade = CascadeType.ALL)
 	@JoinColumn(name = "syllabus_id", referencedColumnName = "sid")
 	private Syllabus syllabus;
-	
+
 	@OneToMany(mappedBy = "course", cascade = CascadeType.ALL)
 	private List<Batch> batches;
-	
+
 	@OneToMany(mappedBy = "course", cascade = CascadeType.ALL)
 	private List<Video> videos;
